Add Image and Youtube components moved to separate files
